MAINT: remove `sparse` dependency from `_lib`
[skip cirrus]
