+++ scipy/optimize/_lsq/common.py
@@ -5,8 +5,8 @@ import numpy as np
 from numpy.linalg import norm
 
 from scipy.linalg import cho_factor, cho_solve, LinAlgError
-from scipy.sparse import issparse
 from scipy.sparse.linalg import LinearOperator, aslinearoperator
+from scipy._lib._sparse import issparse
 
 
 EPS = np.finfo(float).eps
