+++ scipy/optimize/_lsq/least_squares.py
@@ -4,11 +4,11 @@ from warnings import warn
 import numpy as np
 from numpy.linalg import norm
 
-from scipy.sparse import issparse
 from scipy.sparse.linalg import LinearOperator
 from scipy.optimize import _minpack, OptimizeResult
 from scipy.optimize._numdiff import approx_derivative, group_columns
 from scipy.optimize._minimize import Bounds
+from scipy._lib._sparse import issparse
 
 from .trf import trf
 from .dogbox import dogbox
